+++ mantaray/node_test.go
@@ -7,7 +7,6 @@ package mantaray
 import (
 	"bytes"
 	"errors"
-	"fmt"
 	"testing"
 )
 
@@ -118,141 +117,6 @@ func TestRemove(t *testing.T) {
 				if !errors.Is(err, ErrNotFound) {
 					t.Fatalf("expected not found error, got %v", err)
 				}
-			}
-
-		})
-	}
-}
-
-func TestWalkNode(t *testing.T) {
-	for _, tc := range []struct {
-		name  string
-		toAdd [][]byte
-	}{
-		{
-			name: "simple",
-			toAdd: [][]byte{
-				[]byte("index.html"),
-				[]byte("img/1.png"),
-				[]byte("img/2.png"),
-				[]byte("robots.txt"),
-			},
-		},
-	} {
-		t.Run(tc.name, func(t *testing.T) {
-			n := New()
-
-			for i := 0; i < len(tc.toAdd); i++ {
-				c := tc.toAdd[i]
-				e := append(make([]byte, 32-len(c)), c...)
-				err := n.Add(c, e, nil)
-				if err != nil {
-					t.Fatalf("expected no error, got %v", err)
-				}
-			}
-
-			walkedCount := 0
-
-			walker := func(path []byte, node *Node, err error) error {
-				walkedCount++
-
-				pathFound := false
-
-				for i := 0; i < len(tc.toAdd); i++ {
-					c := tc.toAdd[i]
-					if bytes.Equal(path, c) {
-						pathFound = true
-						break
-					}
-				}
-
-				if !pathFound {
-					return fmt.Errorf("walkFn returned unknown path: %s", path)
-				}
-
-				return nil
-			}
-			// Expect no errors.
-			err := n.WalkNode([]byte{}, nil, walker)
-			if err != nil {
-				t.Fatalf("no error expected, found: %s", err)
-			}
-
-			if len(tc.toAdd) != walkedCount {
-				t.Errorf("expected %d nodes, got %d", len(tc.toAdd), walkedCount)
-			}
-
-		})
-	}
-}
-
-func TestWalk(t *testing.T) {
-	for _, tc := range []struct {
-		name     string
-		toAdd    [][]byte
-		expected [][]byte
-	}{
-		{
-			name: "simple",
-			toAdd: [][]byte{
-				[]byte("index.html"),
-				[]byte("img/test/"),
-				[]byte("img/test/oho.png"),
-				[]byte("img/test/old/test.png"),
-				[]byte("robots.txt"),
-			},
-			expected: [][]byte{
-				[]byte("index.html"),
-				[]byte("img"),
-				[]byte("img/test"),
-				[]byte("img/test/oho.png"),
-				[]byte("img/test/old"),
-				[]byte("img/test/old/test.png"),
-				[]byte("robots.txt"),
-			},
-		},
-	} {
-		t.Run(tc.name, func(t *testing.T) {
-			n := New()
-
-			for i := 0; i < len(tc.toAdd); i++ {
-				c := tc.toAdd[i]
-				e := append(make([]byte, 32-len(c)), c...)
-				err := n.Add(c, e, nil)
-				if err != nil {
-					t.Fatalf("expected no error, got %v", err)
-				}
-			}
-
-			walkedCount := 0
-
-			walker := func(path []byte, isDir bool, err error) error {
-				walkedCount++
-
-				pathFound := false
-
-				for i := 0; i < len(tc.expected); i++ {
-					c := tc.expected[i]
-					if bytes.Equal(path, c) {
-						pathFound = true
-						break
-					}
-				}
-
-				if !pathFound {
-					return fmt.Errorf("walkFn returned unknown path: %s", path)
-				}
-
-				return nil
-			}
-			// Expect no errors.
-			err := n.Walk([]byte{}, nil, walker)
-			if err != nil {
-				t.Fatalf("no error expected, found: %s", err)
-			}
-
-			if len(tc.expected) != walkedCount {
-				t.Errorf("expected %d nodes, got %d", len(tc.expected), walkedCount)
 			}
 
 		})
